Garam Masala Devnet

+++ contracts/interfaces/ICapacitorFactory.sol
@@ -23,7 +23,7 @@ interface ICapacitorFactory {
      */
     function deploy(
         uint256 capacitorType,
-        uint256 siblingChainSlug,
+        uint32 siblingChainSlug,
         uint256 maxPacketLength
     ) external returns (ICapacitor, IDecapacitor);
 }
